app/requirements.txt
click==8.0.3
Flask==2.0.2
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.5
# opencv-python==4.1.2.30
# opencv-python-headless==4.1.2.30
opencv-python-headless==4.3.0.38
# opencv-python==4.3.0.38
Pillow==8.4.0
requests==2.6.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
# gunicorn==20.1.0